+++ src/hexkit/providers/akafka/provider/eventsub.py
@@ -15,8 +15,8 @@
 #
 
 """
-Apache Kafka-specific event publisher and subscriber provider.
-They correspond to the `EventPublisherProtocol` and `EventSubscriberProtocol`, respectively.
+Apache Kafka-specific subscriber provider corresponding to the
+`EventSubscriberProtocol`.
 
 Require dependencies of the `akafka` extra. See the `setup.cfg`.
 """
@@ -25,87 +25,19 @@ import json
 import logging
 import ssl
 from contextlib import asynccontextmanager
-from typing import Any, Callable, Literal, Optional, Protocol, TypeVar
+from typing import Callable, Literal, Optional, Protocol, TypeVar
 
-from aiokafka import AIOKafkaConsumer, AIOKafkaProducer
-from aiokafka.helpers import create_ssl_context
-from pydantic import Field
-from pydantic_settings import BaseSettings
+from aiokafka import AIOKafkaConsumer
 
 from hexkit.base import InboundProviderBase
-from hexkit.correlation import (
-    CorrelationIdContextError,
-    get_correlation_id,
-    new_correlation_id,
-    set_correlation_id,
-    validate_correlation_id,
-)
+from hexkit.correlation import set_correlation_id
 from hexkit.custom_types import Ascii, JsonObject
-from hexkit.protocols.eventpub import EventPublisherProtocol
 from hexkit.protocols.eventsub import EventSubscriberProtocol
-
-__all__ = [
-    "KafkaConfig",
-    "KafkaEventPublisher",
-    "ConsumerEvent",
-    "KafkaEventSubscriber",
-]
-
-
-class KafkaConfig(BaseSettings):
-    """Config parameters needed for connecting to Apache Kafka."""
-
-    service_name: str = Field(
-        ...,
-        examples=["my-cool-special-service"],
-        description="The name of the (micro-)service from which messages are published.",
-    )
-    service_instance_id: str = Field(
-        ...,
-        examples=["germany-bw-instance-001"],
-        description=(
-            "A string that uniquely identifies this instance across all instances of"
-            + " this service. A globally unique Kafka client ID will be created by"
-            + " concatenating the service_name and the service_instance_id."
-        ),
-    )
-    kafka_servers: list[str] = Field(
-        ...,
-        examples=[["localhost:9092"]],
-        description="A list of connection strings to connect to Kafka bootstrap servers.",
-    )
-    kafka_security_protocol: Literal["PLAINTEXT", "SSL"] = Field(
-        default="PLAINTEXT",
-        description="Protocol used to communicate with brokers. "
-        + "Valid values are: PLAINTEXT, SSL.",
-    )
-    kafka_ssl_cafile: str = Field(
-        default="",
-        description="Certificate Authority file path containing certificates"
-        + " used to sign broker certificates. If a CA is not specified, the default"
-        + " system CA will be used if found by OpenSSL.",
-    )
-    kafka_ssl_certfile: str = Field(
-        default="",
-        description="Optional filename of client certificate, as well as any"
-        + " CA certificates needed to establish the certificate's authenticity.",
-    )
-    kafka_ssl_keyfile: str = Field(
-        default="", description="Optional filename containing the client private key."
-    )
-    kafka_ssl_password: str = Field(
-        default="",
-        description="Optional password to be used for the client private key.",
-    )
-    generate_correlation_id: bool = Field(
-        default=True,
-        examples=[True, False],
-        description=(
-            "A flag, which, if False, will result in an error when trying to publish an"
-            + " event without a valid correlation ID set for the context. If True, the a"
-            + " newly correlation ID will be generated and used in the event header."
-        ),
-    )
+from hexkit.providers.akafka.provider.config import KafkaConfig
+from hexkit.providers.akafka.provider.utils import (
+    generate_client_id,
+    generate_ssl_context,
+)
 
 
 class EventHeaderNotFoundError(RuntimeError):
@@ -114,162 +46,6 @@ class EventHeaderNotFoundError(RuntimeError):
     def __init__(self, *, header_name):
         message = f"No '{header_name}' was set in event header."
         super().__init__(message)
-
-
-def generate_client_id(*, service_name: str, instance_id: str) -> str:
-    """
-    Generate client id (from the perspective of the Kafka broker) by concatenating
-    the service name and the client suffix.
-    """
-    return f"{service_name}.{instance_id}"
-
-
-def generate_ssl_context(config: KafkaConfig) -> Optional[ssl.SSLContext]:
-    """Generate SSL context for an encrypted SSL connection to Kafka broker."""
-    return (
-        create_ssl_context(
-            cafile=config.kafka_ssl_cafile,
-            certfile=config.kafka_ssl_certfile,
-            keyfile=config.kafka_ssl_keyfile,
-            password=config.kafka_ssl_password,
-        )
-        if config.kafka_security_protocol == "SSL"
-        else None
-    )
-
-
-class KafkaProducerCompatible(Protocol):
-    """A python duck type protocol describing an AIOKafkaProducer or equivalent."""
-
-    def __init__(  # noqa: PLR0913
-        self,
-        *,
-        bootstrap_servers: str,
-        security_protocol: str,
-        ssl_context: Optional[ssl.SSLContext],
-        client_id: str,
-        key_serializer: Callable[[Any], bytes],
-        value_serializer: Callable[[Any], bytes],
-    ):
-        """
-        Initialize the producer with some config params.
-        Args:
-            bootstrap_servers:
-                Comma-separated list of connection strings pointing to the kafka
-                brokers.
-            client_id:
-                A globally unique ID identifying the Kafka client.
-            key_serializer:
-                Function to serialize the keys into bytes.
-            value_serializer:
-                Function to serialize the values into bytes.
-        """
-        ...
-
-    async def start(self):
-        """Setup the producer."""
-        ...
-
-    async def stop(self):
-        """Teardown the producer."""
-        ...
-
-    async def send_and_wait(self, topic, *, key, value, headers) -> Any:
-        """Send event."""
-        ...
-
-
-class KafkaEventPublisher(EventPublisherProtocol):
-    """Apache Kafka specific event publishing provider."""
-
-    @classmethod
-    @asynccontextmanager
-    async def construct(
-        cls,
-        *,
-        config: KafkaConfig,
-        kafka_producer_cls: type[KafkaProducerCompatible] = AIOKafkaProducer,
-    ):
-        """
-        Setup and teardown KafkaEventPublisher instance with some config params.
-
-        Args:
-            config:
-                Config parameters needed for connecting to Apache Kafka.
-            kafka_producer_cls:
-                Overwrite the used Kafka Producer class. Only intended for unit testing.
-        """
-        client_id = generate_client_id(
-            service_name=config.service_name, instance_id=config.service_instance_id
-        )
-
-        producer = kafka_producer_cls(
-            bootstrap_servers=",".join(config.kafka_servers),
-            security_protocol=config.kafka_security_protocol,
-            ssl_context=generate_ssl_context(config),
-            client_id=client_id,
-            key_serializer=lambda key: key.encode("ascii"),
-            value_serializer=lambda event_value: json.dumps(event_value).encode(
-                "ascii"
-            ),
-        )
-
-        try:
-            await producer.start()
-            yield cls(
-                producer=producer,
-                generate_correlation_id=config.generate_correlation_id,
-            )
-        finally:
-            await producer.stop()
-
-    def __init__(
-        self,
-        *,
-        producer: KafkaProducerCompatible,
-        generate_correlation_id: bool,
-    ):
-        """Please do not call directly! Should be called by the `construct` method.
-        Args:
-            producer:
-                hands over a started AIOKafkaProducer.
-            generate_correlation_id:
-                whether to throw an error or generate a new ID if no correlation ID is
-                set for the context when trying to publish an event. An invalid ID will
-                result in an error in either case.
-        """
-        self._producer = producer
-        self._generate_correlation_id = generate_correlation_id
-
-    async def _publish_validated(
-        self, *, payload: JsonObject, type_: Ascii, key: Ascii, topic: Ascii
-    ) -> None:
-        """Publish an event with already validated topic and type.
-
-        Args:
-            payload (JSON): The payload to ship with the event.
-            type_ (str): The event type. ASCII characters only.
-            key (str): The event type. ASCII characters only.
-            topic (str): The event type. ASCII characters only.
-        """
-        try:
-            correlation_id = get_correlation_id()
-        except CorrelationIdContextError:
-            if not self._generate_correlation_id:
-                raise
-
-            correlation_id = new_correlation_id()
-            logging.info("Generated new correlation ID: %s", correlation_id)
-
-        validate_correlation_id(correlation_id)
-
-        event_headers = [
-            ("type", type_.encode("ascii")),
-            ("correlation_id", correlation_id.encode("ascii")),
-        ]
-        await self._producer.send_and_wait(
-            topic, key=key, value=payload, headers=event_headers
-        )
 
 
 class ConsumerEvent(Protocol):
