+++ tests/integration/test_akafka_testutils.py
@@ -23,8 +23,8 @@ from kafka import KafkaAdminClient, KafkaConsumer, TopicPartition
 from kafka.admin.new_topic import NewTopic
 
 from hexkit.custom_types import Ascii, JsonObject
+from hexkit.protocols.eventsub import EventSubscriberProtocol
 from hexkit.providers.akafka.provider import (
-    EventSubscriberProtocol,
     KafkaConfig,
     KafkaEventPublisher,
     KafkaEventSubscriber,
